requirements.txt
torch
torchvision
scipy
ffmpeg-python